test: equality tests
